requirements.txt
addict
pillow>=8.2.0
matplotlib>=3
numpy>1.15
pandas>=1.0
pyyaml>=5.4.1
scikit-learn>=0.21
tqdm
